welcome/workshop/part-2-sidecar/sidecar-producer/requirements.txt
confluent_kafka
flask_restful
flask
fastavro
python-consul